Move meteor cdp methods
