+++ imports/api/ledger/server/methods.js
@@ -60,56 +60,6 @@ Meteor.methods({
         let validator = Validators.findOne({ delegator_address: address })
         return validator;
     },
-    'cdp.getCDPParams': function () {
-        this.unblock();
-        let url = LCD + '/cdp/parameters';
-        let cdpParams = {};
-
-        try {
-            let response = HTTP.get(url);
-            if (response.statusCode == 200) {
-                cdpParams = JSON.parse(response.content).result;
-                return cdpParams
-            }
-        }
-        catch (e) {
-            console.log(e)
-        }
-    },
-
-    'cdp.getCDPPrice': function (market) {
-        this.unblock();
-        let url = LCD + '/pricefeed/price/' + market;
-        let cdpPrice = null;
-
-        try {
-            let response = HTTP.get(url);
-            if (response.statusCode == 200) {
-                cdpPrice = JSON.parse(response.content).result.price;
-                return cdpPrice
-            }
-        }
-        catch (e) {
-            console.log(e)
-        }
-    },
-
-    'cdp.getDeposits': function (address, collateral) {
-        this.unblock();
-        let url = LCD + '/cdp/cdps/cdp/deposits/' + address + '/' + collateral;
-
-        try {
-            let response = HTTP.get(url);
-            if (response.statusCode == 200) {
-                return JSON.parse(response.content).result
-            }
-        }
-        catch (e) {
-            console.log(url);
-            console.log(e)
-        }
-    },
-
     'account.getIncentive': function () {
         this.unblock();
         let url = LCD + '/incentive/rewards'
